+++ go/CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Use correct default base url
+
 ## [go/v0.1.1] — 2026-05-03
 
 ### Added
@@ -15,5 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [go/v0.1.0] — 2026-05-02
 
+[unreleased]: https://github.com/mantyx-io/mantyx-sdk/compare/go/v0.1.1..HEAD
 [go/v0.1.1]: https://github.com/mantyx-io/mantyx-sdk/compare/go/v0.1.0..go/v0.1.1
 [go/v0.1.0]: https://github.com/mantyx-io/mantyx-sdk/releases/tag/go/v0.1.0