+++ python/CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Use correct default base url
+
 ## [python/v0.1.1] — 2026-05-03
 
 ### Added
@@ -13,4 +19,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changelog generation
 
+[unreleased]: https://github.com/mantyx-io/mantyx-sdk/compare/python/v0.1.1..HEAD
 [python/v0.1.1]: https://github.com/mantyx-io/mantyx-sdk/releases/tag/python/v0.1.1